Time: 119 ms (85.02%) | Memory: 62.8 MB (80.62%) - LeetSync

+++ 2616-maximal-score-after-applying-k-operations/maximal-score-after-applying-k-operations.java
@@ -10,7 +10,7 @@ class Solution {
             queue.add(nums[i]);
         }
         double temp = 0;
-        long poppedValue = 0;
+        int  poppedValue = 0;
        for(int i=1; i<=k; i++){
         // System.out.println("queue.poll" + queue.poll());
         poppedValue =  queue.poll();
